refactor(p2p): Split replication.rs and coordinator.rs (Wave 4)
Split 2 large files (~2.4k lines) into 15 smaller focused modules:

sync/replication.rs (821 lines) → sync/replication/ (6 files):
- config.rs: ReplicationConfig struct
- result.rs: ReplicationResult enum
- handlers.rs: Event handlers (handle_block_received, handle_dag_needs_fetch)
- loop_runner.rs: ReplicationLoop struct with run/drain methods
- recovery.rs: recover_unmerged function
- mod.rs: Module exports and tests

sync/coordinator.rs (1,622 lines) → sync/coordinator/ (9 files):
- result_types.rs: SetReplicatorResult, LoadReplicatorsResult
- access.rs: check_access method
- accessors.rs: Simple getter methods
- broadcast.rs: broadcast_local_update, push_to_replicators
- subscriptions.rs: subscribe/unsubscribe collection/document methods
- replicators.rs: Replicator management methods
- constructor.rs: new() and with_* builder methods
- event_handler.rs: handle_host_event and all event type handlers
- mod.rs: SyncCoordinator struct definition and exports

All 17 unit tests pass. Public API unchanged.
