+++ crates/p2p/src/sync/coordinator/mod.rs
@@ -1,0 +1,108 @@
+//! Sync coordinator for DefraDB P2P synchronization.
+//!
+//! The coordinator ties together:
+//! - P2P host for network communication
+//! - SyncManager for block storage and merge tracking
+//! - Broadcaster for publishing updates
+//!
+//! # Architecture
+//!
+//! ```text
+//! Database Layer
+//!       ↓
+//! SyncCoordinator
+//!       ├── Broadcaster (publish updates)
+//!       ├── SyncManager (store blocks, emit events)
+//!       └── Event loop (receive GossipSub messages)
+//!       ↓
+//! P2PHost (network)
+//! ```
+//!
+//! # Security Model: Two-Level Access Control
+//!
+//! The P2P sync layer implements **collection-level** access control only.
+//! **Document-level** ACP is the responsibility of the database merge layer.
+//!
+//! ## Collection-Level (P2P Layer)
+//!
+//! - Enforced via `check_access()` before processing any sync message
+//! - A peer must be registered as a replicator for a collection
+//! - Unauthorized peers cannot push documents to collections they don't replicate
+//!
+//! ## Document-Level (Database Merge Layer)
+//!
+//! - The P2P layer provides creator/doc_id/collection_id in `SyncEvent::BlockReceived`
+//! - The database merge handler should:
+//!   1. Identify the creator's DID (from the signed block or peer mapping)
+//!   2. Check if the creator has UPDATE permission on the document
+//!   3. If permission denied, log and skip the merge (don't crash)
+//!
+//! This two-level model allows:
+//! - Fast collection-level filtering at the network layer
+//! - Fine-grained document-level checks at the merge layer
+//! - CRDT convergence (eventually consistent merge, possibly with rejected updates)
+
+mod access;
+mod accessors;
+mod broadcast;
+mod constructor;
+mod event_handler;
+mod replicators;
+mod result_types;
+mod subscriptions;
+
+pub use result_types::{LoadReplicatorsResult, SetReplicatorResult};
+
+use std::sync::Arc;
+
+use blockstore::Blockstore;
+
+use crate::bitswap::{AccessMode, ReplicatorRegistry};
+use crate::host::P2PHostHandle;
+
+use super::broadcaster::Broadcaster;
+use super::collection_store::P2PCollectionStorage;
+use super::head_provider::DocumentHeadProvider;
+use super::manager::SyncManager;
+use super::peer_state::PeerStateTracker;
+
+/// Coordinator for P2P synchronization.
+///
+/// This is the main integration point between the P2P layer and the database.
+pub struct SyncCoordinator<B: Blockstore> {
+    /// Host handle for sending responses
+    pub(super) host: P2PHostHandle,
+
+    /// Broadcaster for publishing updates
+    pub(super) broadcaster: Broadcaster,
+
+    /// Sync manager for block storage
+    pub(super) manager: SyncManager<B>,
+
+    /// Peer state tracker
+    pub(super) peer_state: Arc<PeerStateTracker>,
+
+    /// Local peer ID (for creator field in broadcasts)
+    pub(super) local_peer_id: String,
+
+    /// Access control mode
+    pub(super) access_mode: AccessMode,
+
+    /// Replicator registry for access control checks
+    pub(super) replicators: Arc<ReplicatorRegistry>,
+
+    /// Set of subscribed collection IDs for P2P sync (in-memory cache)
+    pub(super) subscribed_collections: Arc<tokio::sync::RwLock<std::collections::HashSet<String>>>,
+
+    /// Persistent storage for P2P collection subscriptions
+    pub(super) collection_store: Arc<dyn P2PCollectionStorage>,
+
+    /// Document head provider for DocSync responses
+    pub(super) head_provider: Arc<dyn DocumentHeadProvider>,
+}
+
+#[cfg(test)]
+mod tests {
+    // Integration tests are in tests/integration.rs
+    // Unit tests for individual components are in their respective modules
+}
